# docs/openapi/pull-signal_v1.yaml
openapi: 3.0.3
info:
  title: Interoperability API Gateway Micro Service
  description: exposes the API for interacting with interoperability features
  version: '1.0'
  contact:
    name: API Support
    url: 'http://www.example.com/support'
    email: support@example.com
  termsOfService: 'http://swagger.io/terms/'
  x-api-id: an x-api-id
  x-summary: an x-summary
servers:
  - url: '/api-gateway/1.0'
    description: This service is the interoperability API gateway
tags:
  - name: gateway
    description: Everything about your APIs
    externalDocs:
      description: Find out more
      url: 'http://swagger.io'
  - name: health
    description: Verify service status
    externalDocs:
      description: Find out more
      url: http://swagger.io
security:
  - bearerAuth: []
paths:
  '/pull-signal/{eserviceId}':
    get:
      tags:
        - gateway
      summary: Get a list of signals
      description: Retrieve a list o signals on a specific eservice starting from signalId
      operationId: getRequest
      security:
        - bearerAuth: []
      parameters:
        - name: eserviceId
          in: path
          description: id of the eservice on witch retrieve signals
          required: true
          schema:
            type: string
        - name: signalId
          in: query
          description: signalId from which to search
          required: true
          schema:
            type: integer
            format: int64
            default: 0
        - name: size
          in: query
          description: number of signals to search
          required: false
          schema:
            type: integer
            format: int64
            default: 0

      responses:
        '200':
          description: Full content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationSignal'
          headers:
            'X-Rate-Limit-Limit':
              schema:
                type: integer
              description: Max allowed requests within time interval
            'X-Rate-Limit-Remaining':
              schema:
                type: integer
              description: Remaining requests within time interval
            'X-Rate-Limit-Interval':
              schema:
                type: integer
              description: Time interval in milliseconds. Allowed requests will be constantly replenished during the interval. At the end of the interval the max allowed requests will be available
            'X-Last-Signal-Index':
              schema:
                type: integer
              description: Last signal index

        '206':
          description: Partial content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationSignal'
          headers:
            'X-Rate-Limit-Limit':
              schema:
                type: integer
              description: Max allowed requests within time interval
            'X-Rate-Limit-Remaining':
              schema:
                type: integer
              description: Remaining requests within time interval
            'X-Rate-Limit-Interval':
              schema:
                type: integer
              description: Time interval in milliseconds. Allowed requests will be constantly replenished during the interval. At the end of the interval the max allowed requests will be available
            'X-Last-Signal-Index':
              schema:
                type: integer
              description: Last signal index
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          headers:
            'X-Rate-Limit-Limit':
              schema:
                type: integer
              description: Max allowed requests within time interval
            'X-Rate-Limit-Remaining':
              schema:
                type: integer
              description: Remaining requests within time interval
            'X-Rate-Limit-Interval':
              schema:
                type: integer
              description: Time interval in milliseconds. Allowed requests will be constantly replenished during the interval. At the end of the interval the max allowed requests will be available
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          headers:
            'X-Rate-Limit-Limit':
              schema:
                type: integer
              description: Max allowed requests within time interval
            'X-Rate-Limit-Remaining':
              schema:
                type: integer
              description: Remaining requests within time interval
            'X-Rate-Limit-Interval':
              schema:
                type: integer
              description: Time interval in milliseconds. Allowed requests will be constantly replenished during the interval. At the end of the interval the max allowed requests will be available
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          headers:
            'X-Rate-Limit-Limit':
              schema:
                type: integer
              description: Max allowed requests within time interval
            'X-Rate-Limit-Remaining':
              schema:
                type: integer
              description: Remaining requests within time interval
            'X-Rate-Limit-Interval':
              schema:
                type: integer
              description: Time interval in milliseconds. Allowed requests will be constantly replenished during the interval. At the end of the interval the max allowed requests will be available
        '404':
          description: Request not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          headers:
            'X-Rate-Limit-Limit':
              schema:
                type: integer
              description: Max allowed requests within time interval
            'X-Rate-Limit-Remaining':
              schema:
                type: integer
              description: Remaining requests within time interval
            'X-Rate-Limit-Interval':
              schema:
                type: integer
              description: Time interval in milliseconds. Allowed requests will be constantly replenished during the interval. At the end of the interval the max allowed requests will be available
        '429':
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          headers:
            'X-Rate-Limit-Limit':
              schema:
                type: integer
              description: Max allowed requests within time interval
            'X-Rate-Limit-Remaining':
              schema:
                type: integer
              description: Remaining requests within time interval
            'X-Rate-Limit-Interval':
              schema:
                type: integer
              description: Time interval in milliseconds. Allowed requests will be constantly replenished during the interval. At the end of the interval the max allowed requests will be available
components:
  schemas:
    Signal:
      type: object
      additionalProperties: false
      title: Signal
      description: Signal
      properties:
        signalType:
          $ref: '#/components/schemas/SignalType'
        objectId:
          type: string
          description: object id
          pattern: '^[a-zA-Z0-9\-_]+$'
          minLength: 0
          maxLength: 500
        objectType:
          type: string
          description: object type
          pattern: '^[a-zA-Z0-9\s\-_]+$'
          minLength: 0
          maxLength: 500
        eserviceId:
          type: string
          description: eservice Id
        signalId:
          type: integer
          description: sorting signal index
          format: int64

    PaginationSignal:
      type: object
      description: PaginationSignal
      properties:
        signals:
          type: array
          items:
            $ref: '#/components/schemas/Signal'
        lastSignalId:
          type: integer
          format: int64

    SignalType:
      type: string
      description: type of signal
      enum:
        - CREATE
        - UPDATE
        - DELETE
    Problem:
      properties:
        type:
          description: URI reference of type definition
          type: string
        status:
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          example: 400
          exclusiveMaximum: true
          format: int32
          maximum: 600
          minimum: 100
          type: integer
        title:
          description: A short, summary of the problem type. Written in english and readable
          example: Service Unavailable
          maxLength: 64
          pattern: '^[ -~]{0,64}$'
          type: string
        correlationId:
          description: Unique identifier of the request
          example: '53af4f2d-0c87-41ef-a645-b726a821852b'
          maxLength: 64
          type: string
        detail:
          description: A human readable explanation of the problem.
          example: Request took too long to complete.
          maxLength: 4096
          pattern: '^.{0,1024}$'
          type: string
        errors:
          type: array
          minItems: 0
          items:
            $ref: '#/components/schemas/ProblemError'
      additionalProperties: false
      required:
        - type
        - status
        - title
        - errors
    ProblemError:
      properties:
        code:
          description: Internal code of the error
          example: 123-4567
          minLength: 8
          maxLength: 8
          pattern: '^[0-9]{3}-[0-9]{4}$'
          type: string
        detail:
          description: A human readable explanation specific to this occurrence of the problem.
          example: Parameter not valid
          maxLength: 4096
          pattern: '^.{0,1024}$'
          type: string
      required:
        - code
        - detail
  securitySchemes:
    bearerAuth:
      type: http
      description: 'A bearer token in the format of a JWS and conformed to the specifications included in [RFC8725](https://tools.ietf.org/html/RFC8725).'
      scheme: bearer
      bearerFormat: JWT